iaGenerativa_taller2/pyproject.toml
[tool.poetry]
name = "iaGenerativa_taller2"
version = "0.1.0"
description = "A project for processing documents and generating embeddings."
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.0"
PyPDF2 = "^1.26.0"
pinecone-client = "^2.0.0"
langchain = "^0.0.1"
langchain-community = "^0.0.1"
rich = "^12.0.0"
fastapi = "^0.95.0"
openpyxl = "^3.0.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"